+++ actix/src/lib.rs
@@ -18,7 +18,7 @@
 //! - Using Futures for asynchronous message handling
 //! - Actor supervision
 //! - Typed messages (No [`Any`](std::any::Any) type). Generic messages are allowed
-//! - Runs on stable Rust 1.65+
+//! - Runs on stable Rust 1.68+
 
 #![deny(rust_2018_idioms, nonstandard_style, future_incompatible)]
 #![doc(html_logo_url = "https://actix.rs/img/logo.png")]