+++ actix-broker/CHANGES.md
@@ -4,9 +4,10 @@
 
 
 ## 0.4.3 - 2022-05-24
-- Fix at-least-once delivery guarantee for IssueAsync. [#536]
+- Fix at-least-once delivery guarantee for `IssueAsync`. [#536]
 
 [#536]: https://github.com/actix/actix/pull/536
+
 
 ## 0.4.2 - 2022-03-01
 - Add support for `actix` v0.13.