+++ actix-broker/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.68.
+- Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 0.4.3 - 2022-05-24