+++ actix-broker/CHANGES.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
+- Improve broker message performance.
 - Minimum supported Rust version (MSRV) is now 1.75.
-- Improve Broker message performance - take_subs() function.
 
 ## 0.4.3 - 2022-05-24