+++ actix-broker/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.4.4
+
 - Improve broker message performance.
 - Minimum supported Rust version (MSRV) is now 1.75.